requirements.txt
# Core bot dependencies
aiogram==3.4.1
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1

# Image processing (headless for Docker, no GUI dependencies)
opencv-python-headless==4.8.1.78
Pillow==10.0.1
numpy==1.24.3

# Utilities
python-dotenv==1.0.0
redis==5.0.1
aiofiles==23.2.1
psutil==5.9.6

# Database driver
psycopg2-binary==2.9.9

# Protobuf compatibility
protobuf>=3.20.0,<4.0.0